Issue #22: improve test coverage for base_http.py
In #19, we introduced some code that was missing test coverage for lines
897, 925, and 943-944. When possible, we wanted to add some unit test
coverage for those lines.

For this commit, we refactored out some of the logic into unit-testable
methods and added unit tests (line 925 and 943-944). For 897, we added a
`pragma: no cover` because AsyncMocks were not introduced until 3.8, and
our current ucsschool system defaults to 3.7, so it did not make sense
to write those tests at this time.